+++ CHANGELOG.md
@@ -28,7 +28,7 @@ Version 7 of the SDK is compatible with Sentry self-hosted versions 24.4.2 or hi
 
 ### Major Changes
 
-- `ip addresses` is only collected when `sendDefaultPii`: `true`
+- Set `{{auto}}` if `user.ip_address` is `undefined` and `sendDefaultPii: true` ([#4466](https://github.com/getsentry/sentry-react-native/pull/4466))
 - Exceptions from `captureConsoleIntegration` are now marked as handled: true by default
 - `shutdownTimeout` moved from `core` to `@sentry/react-native`
 - `hasTracingEnabled` was renamed to `hasSpansEnabled`
@@ -290,7 +290,6 @@ Change `Cold/Warm App Start` span description to `Cold/Warm Start` ([#4636](http
 - Add mechanism field to unhandled rejection errors ([#4457](https://github.com/getsentry/sentry-react-native/pull/4457))
 - Use proper SDK name for Session Replay tags ([#4428](https://github.com/getsentry/sentry-react-native/pull/4428))
 - Use `makeDsn` from `core` to extract the URL from DSN avoiding unimplemented `URL.protocol` errors ([#4395](https://github.com/getsentry/sentry-react-native/pull/4395))
-- Set `{{auto}}` if `user.ip_address` is `undefined` ([#4466](https://github.com/getsentry/sentry-react-native/pull/4466))
 
 ### Changes